Send the mac scripts to build machine

+++ nbbuild/newbuild/build-nbi.sh
@@ -9,6 +9,15 @@ source init.sh
 cd $NB_ALL/installer/infra/build
 
 if [ ! -z $NATIVE_MAC_MACHINE ]; then
+   ssh $NATIVE_MAC_MACHINE rm -rf $MAC_PATH/installer
+   ERROR_CODE=$?
+   if [ $ERROR_CODE != 0 ]; then
+       echo "ERROR: $ERROR_CODE - Connection to MAC machine $NATIVE_MAC_MACHINE failed, can't remove old scripts"
+       exit $ERROR_CODE;
+   fi
+   ssh $NATIVE_MAC_MACHINE mkdir $MAC_PATH/installer
+   cd $NB_ALL
+   gtar c installer/mac | ssh $NATIVE_MAC_MACHINE "( cd $MAC_PATH; tar x )"
    ssh $NATIVE_MAC_MACHINE rm -f $MAC_PATH/zip/*
    ERROR_CODE=$?
    if [ $ERROR_CODE != 0 ]; then
